Update: Checkpoints and cameras
i added a checkpoint system so that if the player falls off in a place they arent meant to, they will be reset based on the last checkpoint that theyve activated. also updates the cameras which had weird perspectives during the gameplay testing. Finally, I have added invisible barriers to areas that needed them to stop being softlocked and/ or skipping main parts of the map.